table adjusted and seeded

+++ db/migrate/20210325180542_create_carts.rb
@@ -2,7 +2,7 @@ class CreateCarts < ActiveRecord::Migration[6.1]
   def change
     create_table :carts do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :total
+      t.float :total
 
       t.timestamps
     end
